code cleanup
blank card added

cards flips
