+++ memory_game/Assets/Scripts/Card.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a1f420d759582f942977201afc10fefe
+timeCreated: 1458904522
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
