+++ memory_game/Assets/Scripts/GameScipt.cs
@@ -1,13 +1,13 @@
 ﻿using UnityEngine;
 using System.Collections;
 using System.Collections.Generic;
-using System;
 using System.Linq;
-using System.Threading;
+using Assets.Scripts;
 
-public class GameScipt : MonoBehaviour {
-
+public class GameScipt : MonoBehaviour
+{
     public GUISkin customSkin;
+
     static int columns = 4;
     static int rows = 4;
     static int totalCards = columns * rows;
@@ -15,11 +15,11 @@ public class GameScipt : MonoBehaviour {
     int machesMade = 0;
     int cardWidth = 100;
     int cardHeight = 100;
+    bool playerCanClick; //flag to prevent clicking
+    bool playerHasWon = false;
     Card[,] gridOfCards;
     ArrayList arrayCardsFlipped;
     List<Card> deckOfCards = new List<Card>();
-    bool playerCanClick; //flag to prevent clicking
-    bool playerHasWon = false;
     List<string> images = new List<string>(new string[] { "elephant", "giraffe", "gorilla", "lion", "moose", "hippopotamus", "sloth", "zebra" });
     Card card;
 
@@ -44,16 +44,15 @@ public class GameScipt : MonoBehaviour {
 
     private void BuildDeck()
     {
-        for(int i =0; i<2; i++)
+        for (int i = 0; i < 2; i++)
         {
             for (int j = 0; j < totalCards / 2; j++)
             {
                 card = new Card(images.ElementAt(j));
                 deckOfCards.Add(card);
-                
             }
         }
-        deckOfCards.Shuffle();       
+        deckOfCards.Shuffle();
     }
 
     void OnGUI()
@@ -74,62 +73,29 @@ public class GameScipt : MonoBehaviour {
             for (int j = 0; j < columns; j++)
             {
                 Card card = gridOfCards[i, j];
-                if (GUILayout.Button(Resources.Load(card.img) as Texture2D, GUILayout.Width(cardWidth)))
+                string img;
+                if (card.isFaceUp)
                 {
-                    Debug.Log(card.img);
+                    img = card.img;
+                }
+                else
+                {
+                    img = card.imgDown;
+                }
+                if (GUILayout.Button(Resources.Load(img) as Texture2D, GUILayout.Width(cardWidth)))
+                {
+                    FlipCardFaceUp(card);
                 }
             }
             GUILayout.FlexibleSpace();
             GUILayout.EndHorizontal();
         }
         GUILayout.FlexibleSpace();
-
         GUILayout.EndVertical();
     }
 
-    
-   
-    public class Card : System.Object
+    private void FlipCardFaceUp(Card card)
     {
-        public bool isFaceUp = false;
-        public bool isMatched = false;
-        public string img;
-
-        public Card()
-        {
-            img = "lion";
-        }
-
-        public Card(string img)
-        {
-            this.img = img;
-        }
-    }
-}
-
-public static class ThreadSafeRandom
-{
-    [ThreadStatic]
-    private static System.Random Local;
-
-    public static System.Random ThisThreadsRandom
-    {
-        get { return Local ?? (Local = new System.Random(unchecked(Environment.TickCount * 31 + Thread.CurrentThread.ManagedThreadId))); }
-    }
-}
-
-static class MyExtensions
-{
-        public static void Shuffle<T>(this IList<T> list)
-    {
-        int n = list.Count;
-        while (n > 1)
-        {
-            n--;
-            int k = ThreadSafeRandom.ThisThreadsRandom.Next(n + 1);
-            T value = list[k];
-            list[k] = list[n];
-            list[n] = value;
-        }
+        card.isFaceUp = true;
     }
 }
