+++ memory_game/Assets/Scripts/MyExtension.cs
@@ -1,0 +1,20 @@
+﻿using System.Collections.Generic;
+
+namespace Assets.Scripts
+{
+    static class MyExtensions
+    {
+        public static void Shuffle<T>(this IList<T> list)
+        {
+            int n = list.Count;
+            while (n > 1)
+            {
+                n--;
+                int k = ThreadSafeRandom.ThisThreadsRandom.Next(n + 1);
+                T value = list[k];
+                list[k] = list[n];
+                list[n] = value;
+            }
+        }
+    }
+}
