+++ memory_game/Assets/Scripts/ThreadSafeRandom.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Threading;
+
+namespace Assets.Scripts
+{
+    public static class ThreadSafeRandom
+    {
+        [ThreadStatic]
+        private static System.Random Local;
+
+        public static System.Random ThisThreadsRandom
+        {
+            get { return Local ?? (Local = new System.Random(unchecked(Environment.TickCount * 31 + Thread.CurrentThread.ManagedThreadId))); }
+        }
+    }
+}
