+++ memory_game/memory_game.sln
@@ -1,7 +1,7 @@
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2008
-
-Project("{53A001EA-452A-208E-3FCF-21E98138A670}") = "memory_game", "Assembly-CSharp.csproj", "{1AB87CDA-C381-7D59-79F5-7A97E22BFC19}"
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2015
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "memory_game.CSharp", "memory_game.CSharp.csproj", "{F1CB8681-C0D4-B31B-C970-80D6737EC9E0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -9,15 +9,12 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1AB87CDA-C381-7D59-79F5-7A97E22BFC19}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1AB87CDA-C381-7D59-79F5-7A97E22BFC19}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1AB87CDA-C381-7D59-79F5-7A97E22BFC19}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1AB87CDA-C381-7D59-79F5-7A97E22BFC19}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F1CB8681-C0D4-B31B-C970-80D6737EC9E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F1CB8681-C0D4-B31B-C970-80D6737EC9E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F1CB8681-C0D4-B31B-C970-80D6737EC9E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F1CB8681-C0D4-B31B-C970-80D6737EC9E0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		StartupItem = Assembly-CSharp.csproj
 	EndGlobalSection
 EndGlobal
